--- src/main/java/study/refactoring/chapter1/play/Play.java
@@ -1,4 +1,0 @@
-package study.refactoring.chapter1.play;
-
-public record Play(String name, PlayType type) {
-}
