+++ src/main/java/study/refactoring/chapter1/play/dto/Play.java
@@ -1,0 +1,6 @@
+package study.refactoring.chapter1.play.dto;
+
+import study.refactoring.chapter1.play.PlayType;
+
+public record Play(String name, PlayType type) {
+}
