+++ src/main/java/study/refactoring/chapter1/statement/dto/EnrichedPerformance.java
@@ -1,6 +1,6 @@
-package study.refactoring.chapter1.statement;
+package study.refactoring.chapter1.statement.dto;
 
-import study.refactoring.chapter1.play.Play;
+import study.refactoring.chapter1.play.dto.Play;
 
 public record EnrichedPerformance(String playID, int audience, Play play, double amount, int volumeCredits) {
 }
