+++ src/main/java/study/refactoring/chapter1/statement/dto/Performance.java
@@ -1,4 +1,4 @@
-package study.refactoring.chapter1.statement;
+package study.refactoring.chapter1.statement.dto;
 
 public record Performance(String playID, int audience) {
 }
